+++ schug/load/biomart.py
@@ -9,7 +9,7 @@ import requests
 
 LOG = logging.getLogger(__name__)
 BIOMART_37_URL = "https://grch37.ensembl.org/biomart/martservice/?query="
-BIOMART_38_URL = "https://www.ensembl.org/biomart/martservice/?query="
+BIOMART_38_URL = "https://jun2026.archive.ensembl.org/biomart/martservice/?query="
 
 
 class EnsemblOutageError(Exception):